+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+- Improve Converter template.py usability
+
 ## [v0.10.0] - 2025-03-11
 
 ### Added